+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/transaction/impl/expectations/TrackingKeyValueServiceImpl.java
@@ -53,7 +53,7 @@ public final class TrackingKeyValueServiceImpl extends ForwardingKeyValueService
     private static final SafeLogger log = SafeLoggerFactory.get(TrackingKeyValueServiceImpl.class);
 
     private final KeyValueService delegate;
-    private final KeyValueServiceDataTracker tracker = KeyValueServiceDataTrackerImpl.createFailSafe();
+    private final KeyValueServiceDataTracker tracker = new KeyValueServiceDataTracker();
 
     public TrackingKeyValueServiceImpl(KeyValueService delegate) {
         this.delegate = delegate;
